fix client bundle crash on /owner and /criar-area
The "use client" CreatorAreaCreateForm imported the creator-area zod schema
and error formatters from creators/service.ts, which imports getDb from
db/client, which imports @/lib/env. env.ts throws at module evaluation when
NEXTAUTH_SECRET is missing in production. Bundled into the client, that check
runs in the browser (where the secret is absent) and crashes hydration, so
both pages white-screened with "This page couldn't load" and no server logs.

Extract the client-safe validation/formatting into creators/area-form.ts (no
db/env imports), point the form at it, and re-export from service.ts for
server callers. Also scope env.ts's production throw to the server
(typeof window === "undefined") as defense-in-depth.
