+++ src/lib/creators/service.ts
@@ -1,8 +1,13 @@
 import { randomUUID } from "node:crypto";
 
 import { desc, eq } from "drizzle-orm";
-import { z, ZodError } from "zod";
 
+import {
+  createCreatorAreaSchema,
+  flattenCreatorAreaSchemaErrors,
+  formatCreateCreatorAreaError,
+  type CreateCreatorAreaInput,
+} from "@/lib/creators/area-form";
 import { DEFAULT_CREATOR_BRANDING, DEFAULT_CREATOR_DOMAIN } from "@/lib/creators/defaults";
 import {
   buildDemoCreatorModules,
@@ -17,19 +22,16 @@ import { getDb } from "@/lib/db/client";
 import { creatorBranding, creatorDomains, creatorModules, creators } from "@/lib/db/schema";
 import type { CreatorRecord, CreatorTenantRecord } from "@/lib/types";
 
-const creatorColorSchema = z
-  .string()
-  .trim()
-  .regex(/^#[0-9a-fA-F]{6}$/u, "Use uma cor em hexadecimal, como #c7a2e9.");
-
-export const createCreatorAreaSchema = z.object({
-  displayName: z.string().trim().min(2, "Informe o nome do criador.").max(80, "Use até 80 caracteres."),
-  slug: z.string().trim().max(64, "Use até 64 caracteres.").optional(),
-  primaryColor: creatorColorSchema.default(DEFAULT_CREATOR_BRANDING.primaryColor),
-  accentColor: creatorColorSchema.default(DEFAULT_CREATOR_BRANDING.accentColor),
-});
-
-export type CreateCreatorAreaInput = z.infer<typeof createCreatorAreaSchema>;
+// Client-safe validation/formatting lives in area-form.ts so the "use client"
+// CreatorAreaCreateForm can import it without dragging this server module (and
+// its db/env dependencies) into the browser bundle. Re-exported here so
+// existing server-side imports from "@/lib/creators/service" keep working.
+export {
+  createCreatorAreaSchema,
+  formatCreateCreatorAreaError,
+  flattenCreatorAreaSchemaErrors,
+  type CreateCreatorAreaInput,
+};
 
 export type CreatorAreaSummary = CreatorRecord & {
   publicPath: string;
@@ -62,10 +64,6 @@ function toAreaSummary(creator: CreatorRecord): CreatorAreaSummary {
     publicPath: `/c/${creator.slug}`,
     publicHostname: `${creator.slug}.${DEFAULT_CREATOR_DOMAIN}`,
   };
-}
-
-function formatCreatorAreaSchemaError(error: ZodError) {
-  return error.issues[0]?.message ?? "Dados inválidos.";
 }
 
 function isMissingCreatorSchemaError(error: unknown) {
@@ -113,37 +111,6 @@ function isMissingCreatorSchemaError(error: unknown) {
   return false;
 }
 
-export function formatCreateCreatorAreaError(error: unknown) {
-  if (error instanceof ZodError) {
-    return formatCreatorAreaSchemaError(error);
-  }
-
-  const message = error instanceof Error ? error.message : "Falha ao criar área.";
-  switch (message) {
-    case "creator_slug_exists":
-      return "Esse endereço já está em uso.";
-    case "creator_slug_reserved":
-      return "Esse endereço é reservado.";
-    case "invalid_creator_slug":
-      return "Use um endereço com letras, números e hífens.";
-    case "missing_creator_owner":
-      return "Entre novamente para criar a área.";
-    case "creator_schema_missing":
-      return "A estrutura de criadores ainda não foi aplicada no banco. Rode as migrações antes de criar áreas.";
-    default:
-      return message;
-  }
-}
-
-export function flattenCreatorAreaSchemaErrors(error: ZodError) {
-  return error.issues.reduce<Partial<Record<keyof CreateCreatorAreaInput, string>>>((acc, issue) => {
-    const field = issue.path[0];
-    if (typeof field === "string" && !acc[field as keyof CreateCreatorAreaInput]) {
-      acc[field as keyof CreateCreatorAreaInput] = issue.message;
-    }
-    return acc;
-  }, {});
-}
 
 function parseCreatorAreaInput(input: unknown) {
   const parsed = createCreatorAreaSchema.parse(input);
